Added GuideDetails.tsx: restruct file
